Move parallel write test to async test file
